simplified the tests by removing the refactored changes
