+++ encoding/thrift/encodingIntegration_test.go
@@ -2,13 +2,17 @@ package thrift_test
 
 import (
 	"context"
+	"net"
 	"testing"
 	"time"
 
+	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	"go.uber.org/yarpc"
+	"go.uber.org/yarpc/api/transport"
 	"go.uber.org/yarpc/encoding/thrift/internal/observabilitytest/test"
 	"go.uber.org/yarpc/encoding/thrift/internal/observabilitytest/test/testserviceclient"
+	"go.uber.org/yarpc/encoding/thrift/internal/observabilitytest/test/testserviceserver"
 	"go.uber.org/yarpc/transport/http"
 	"go.uber.org/yarpc/transport/tchannel"
 	"go.uber.org/yarpc/yarpctest"
@@ -70,7 +74,7 @@ func TestThriftMetrics1(t *testing.T) {
 
 	for _, trans := range transports {
 		t.Run(trans+" thift call", func(t *testing.T) {
-			client, _, _, _, cleanup := initClientAndServer(t, trans, testServer1{})
+			client, cleanup := CreateClientAndServer(t, trans, testServer1{})
 			defer cleanup()
 
 			for _, test := range tests {
@@ -80,6 +84,91 @@ func TestThriftMetrics1(t *testing.T) {
 			}
 		})
 	}
+}
+
+func CreateClientAndServer(
+	t *testing.T,
+	trans string,
+	server testserviceserver.Interface,
+) (
+	client testserviceclient.Interface,
+	cleanup func(),
+) {
+	serverAddr, cleanupServer := CreateServer(t, trans, server)
+	client, cleanupClient := CreateClient(t, trans, serverAddr)
+
+	return client, func() {
+		cleanupServer()
+		cleanupClient()
+	}
+}
+
+func CreateServer(t *testing.T, transportType string, server testserviceserver.Interface) (addr string, cleanup func()) {
+	var inbound transport.Inbound
+
+	switch transportType {
+	case tchannel.TransportName:
+		listen, err := net.Listen("tcp", "127.0.0.1:0")
+		require.NoError(t, err)
+
+		trans, err := tchannel.NewTransport(
+			tchannel.ServiceName(_serverName),
+			tchannel.Listener(listen))
+		require.NoError(t, err)
+
+		inbound = trans.NewInbound()
+		addr = listen.Addr().String()
+
+	case http.TransportName:
+		hInbound := http.NewTransport().NewInbound("127.0.0.1:0")
+		defer func() { addr = "http://" + hInbound.Addr().String() }() // can only get addr after dispatcher has started
+		inbound = hInbound
+
+	default:
+		t.Fatal("unknown transport")
+	}
+
+	dispatcher := yarpc.NewDispatcher(yarpc.Config{
+		Name:     _serverName,
+		Inbounds: yarpc.Inbounds{inbound},
+	})
+
+	dispatcher.Register(testserviceserver.New(server))
+
+	require.NoError(t, dispatcher.Start(), "could not start server dispatcher")
+
+	cleanup = func() { assert.NoError(t, dispatcher.Stop(), "could not stop dispatcher") }
+	return addr, cleanup
+}
+
+func CreateClient(t *testing.T, transportType string, serverAddr string) (client testserviceclient.Interface, cleanup func()) {
+	var out transport.UnaryOutbound
+
+	switch transportType {
+	case tchannel.TransportName:
+		trans, err := tchannel.NewTransport(tchannel.ServiceName(_clientName))
+		require.NoError(t, err)
+		out = trans.NewSingleOutbound(serverAddr)
+
+	case http.TransportName:
+		out = http.NewTransport().NewSingleOutbound(serverAddr)
+	}
+
+	dispatcher := yarpc.NewDispatcher(yarpc.Config{
+		Name: _clientName,
+		Outbounds: map[string]transport.Outbounds{
+			_serverName: {
+				ServiceName: _serverName,
+				Unary:       out,
+			},
+		},
+	})
+
+	client = testserviceclient.New(dispatcher.ClientConfig(_serverName))
+	require.NoError(t, dispatcher.Start(), "could not start client dispatcher")
+
+	cleanup = func() { assert.NoError(t, dispatcher.Stop(), "could not stop dispatcher") }
+	return client, cleanup
 }
 
 type testServer1 struct{}
