+++ encoding/thrift/observability_test.go
@@ -22,18 +22,24 @@ package thrift_test
 
 import (
 	"context"
+	"net"
 	"testing"
 	"time"
 
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
+	"go.uber.org/net/metrics"
+	"go.uber.org/yarpc"
+	"go.uber.org/yarpc/api/transport"
 	"go.uber.org/yarpc/encoding/thrift/internal/observabilitytest/test"
+	"go.uber.org/yarpc/encoding/thrift/internal/observabilitytest/test/testserviceclient"
+	"go.uber.org/yarpc/encoding/thrift/internal/observabilitytest/test/testserviceserver"
 	"go.uber.org/yarpc/internal/testutils"
 	"go.uber.org/yarpc/transport/http"
 	"go.uber.org/yarpc/transport/tchannel"
-	"go.uber.org/yarpc/yarpctest"
 	"go.uber.org/zap"
 	"go.uber.org/zap/zapcore"
+	"go.uber.org/zap/zaptest/observer"
 )
 
 const (
@@ -54,7 +60,7 @@ func TestThriftExceptionObservability(t *testing.T) {
 
 	for _, trans := range transports {
 		t.Run("exception with annotation", func(t *testing.T) {
-			client, observedLogs, clientMetricsRoot, serverMetricsRoot, cleanup := initClientAndServer(t, trans, testServer{})
+			client, observedLogs, clientMetricsRoot, serverMetricsRoot, cleanup := initClientAndServer(t, trans)
 			defer cleanup()
 
 			ctx, cancel := context.WithTimeout(context.Background(), time.Second)
@@ -99,7 +105,7 @@ func TestThriftExceptionObservability(t *testing.T) {
 		})
 
 		t.Run("exception without annotation ", func(t *testing.T) {
-			client, observedLogs, clientMetricsRoot, serverMetricsRoot, cleanup := initClientAndServer(t, trans, testServer{})
+			client, observedLogs, clientMetricsRoot, serverMetricsRoot, cleanup := initClientAndServer(t, trans)
 			defer cleanup()
 
 			ctx, cancel := context.WithTimeout(context.Background(), time.Second)
@@ -147,13 +153,10 @@ func TestThriftMetrics(t *testing.T) {
 
 	for _, trans := range transports {
 		t.Run(trans+" thift call", func(t *testing.T) {
-			client, _, clientMetricsRoot, serverMetricsRoot, cleanup := initClientAndServer(t, trans, testServer{})
+			client, _, clientMetricsRoot, serverMetricsRoot, cleanup := initClientAndServer(t, trans)
 			defer cleanup()
 
 			ctx, cancel := context.WithTimeout(context.Background(), time.Second)
-			ctx = yarpctest.ContextWithCall(ctx, &yarpctest.Call{
-				Procedure: "ABC1",
-			})
 			defer cancel()
 
 			_, err := client.Call(ctx, _wantSuccess)
@@ -181,6 +184,137 @@ func TestThriftMetrics(t *testing.T) {
 			})
 		})
 	}
+}
+
+func assertLogs(t *testing.T, wantFields []zapcore.Field, logs []observer.LoggedEntry) {
+	require.Len(t, logs, 2, "unexpected number of logs")
+
+	t.Run("inbound", func(t *testing.T) {
+		require.Equal(t, _errorInbound, logs[0].Message, "unexpected log")
+		assertLogFields(t, wantFields, logs[0].Context)
+	})
+
+	t.Run("outbound", func(t *testing.T) {
+		require.Equal(t, _errorOutbound, logs[1].Message, "unexpected log")
+		assertLogFields(t, wantFields, logs[1].Context)
+	})
+}
+
+func assertLogFields(t *testing.T, wantFields, gotContext []zapcore.Field) {
+	gotFields := make(map[string]zapcore.Field)
+	for _, log := range gotContext {
+		gotFields[log.Key] = log
+	}
+
+	for _, want := range wantFields {
+		got, ok := gotFields[want.Key]
+		if assert.True(t, ok, "key %q not found", want.Key) {
+			assert.Equal(t, want, got, "unexpected log field")
+		}
+	}
+}
+
+func initClientAndServer(
+	t *testing.T,
+	trans string,
+) (
+	client testserviceclient.Interface,
+	observedLogs *observer.ObservedLogs,
+	clientMetricsRoot *metrics.Root,
+	serverMetricsRoot *metrics.Root,
+	cleanup func(),
+) {
+	loggerCore, observedLogs := observer.New(zapcore.DebugLevel)
+	clientMetricsRoot, serverMetricsRoot = metrics.New(), metrics.New()
+
+	serverAddr, cleanupServer := newServer(t, trans, loggerCore, serverMetricsRoot)
+	client, cleanupClient := newClient(t, trans, serverAddr, loggerCore, clientMetricsRoot)
+
+	_ = observedLogs.TakeAll() // ignore all start up logs
+
+	return client, observedLogs, clientMetricsRoot, serverMetricsRoot, func() {
+		cleanupServer()
+		cleanupClient()
+	}
+}
+
+func newServer(t *testing.T, transportType string, loggerCore zapcore.Core, metricsRoot *metrics.Root) (addr string, cleanup func()) {
+	var inbound transport.Inbound
+
+	switch transportType {
+	case tchannel.TransportName:
+		listen, err := net.Listen("tcp", "127.0.0.1:0")
+		require.NoError(t, err)
+
+		trans, err := tchannel.NewTransport(
+			tchannel.ServiceName(_serverName),
+			tchannel.Listener(listen))
+		require.NoError(t, err)
+
+		inbound = trans.NewInbound()
+		addr = listen.Addr().String()
+
+	case http.TransportName:
+		hInbound := http.NewTransport().NewInbound("127.0.0.1:0")
+		defer func() { addr = "http://" + hInbound.Addr().String() }() // can only get addr after dispatcher has started
+		inbound = hInbound
+
+	default:
+		t.Fatal("unknown transport")
+	}
+
+	dispatcher := yarpc.NewDispatcher(yarpc.Config{
+		Name:     _serverName,
+		Inbounds: yarpc.Inbounds{inbound},
+		Logging: yarpc.LoggingConfig{
+			Zap: zap.New(loggerCore),
+		},
+		Metrics: yarpc.MetricsConfig{
+			Metrics: metricsRoot.Scope(),
+		},
+	})
+
+	dispatcher.Register(testserviceserver.New(&testServer{}))
+	require.NoError(t, dispatcher.Start(), "could not start server dispatcher")
+
+	cleanup = func() { assert.NoError(t, dispatcher.Stop(), "could not stop dispatcher") }
+	return addr, cleanup
+}
+
+func newClient(t *testing.T, transportType string, serverAddr string, loggerCore zapcore.Core, metricsRoot *metrics.Root) (client testserviceclient.Interface, cleanup func()) {
+	var out transport.UnaryOutbound
+
+	switch transportType {
+	case tchannel.TransportName:
+		trans, err := tchannel.NewTransport(tchannel.ServiceName(_clientName))
+		require.NoError(t, err)
+		out = trans.NewSingleOutbound(serverAddr)
+
+	case http.TransportName:
+		out = http.NewTransport().NewSingleOutbound(serverAddr)
+	}
+
+	dispatcher := yarpc.NewDispatcher(yarpc.Config{
+		Name: _clientName,
+		Outbounds: map[string]transport.Outbounds{
+			_serverName: {
+				ServiceName: _serverName,
+				Unary:       out,
+			},
+		},
+		Logging: yarpc.LoggingConfig{
+			Zap: zap.New(loggerCore),
+		},
+		Metrics: yarpc.MetricsConfig{
+			Metrics: metricsRoot.Scope(),
+		},
+	})
+
+	client = testserviceclient.New(dispatcher.ClientConfig(_serverName))
+	require.NoError(t, dispatcher.Start(), "could not start client dispatcher")
+
+	cleanup = func() { assert.NoError(t, dispatcher.Stop(), "could not stop dispatcher") }
+	return client, cleanup
 }
 
 type testServer struct{}
